Improve error handling, logging and spans (#15)
Introduces structured logging (JSON) where log messages are tagged with relevant attributes that can be expanded in observability and are visible in stdout.

Refactor error handling so that it all happens in the same asynchronous context where appropriate, rather than mixing synchronous and asynchronous error handling.

Adds a style-guide for logging.